internal/controllers/controllers.go
package controllers

import (
	"context"
	"encoding/json"
	"fmt"

	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
	"k8s.io/apimachinery/pkg/runtime"
	"k8s.io/apimachinery/pkg/runtime/schema"
	"k8s.io/apimachinery/pkg/types"
	"sigs.k8s.io/controller-runtime/pkg/client"
	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"

	corev1alpha1 "package-operator.run/apis/core/v1alpha1"
)

const (
	// This label is set on all dynamic objects to limit caches.
	DynamicCacheLabel = "package-operator.run/cache"
	// Common finalizer to free allocated caches when objects are deleted.
	CachedFinalizer = "package-operator.run/cached"
	// OwnerReference Controller gk:name/namespace local field index.
	ControllerIndexFieldKey = "package-operator.run/controller"
)

// Ensures the given finalizer is set and persisted on the given object.
func EnsureFinalizer(
	ctx context.Context, c client.Client,
	obj client.Object, finalizer string,
) error {
	if controllerutil.ContainsFinalizer(obj, finalizer) {
		return nil
	}

	controllerutil.AddFinalizer(obj, finalizer)
	patch := map[string]interface{}{
		"metadata": map[string]interface{}{
			"resourceVersion": obj.GetResourceVersion(),
			"finalizers":      obj.GetFinalizers(),
		},
	}
	patchJSON, err := json.Marshal(patch)
	if err != nil {
		return fmt.Errorf("marshalling patch to remove finalizer: %w", err)
	}

	if err := c.Patch(ctx, obj, client.RawPatch(types.MergePatchType, patchJSON)); err != nil {
		return fmt.Errorf("adding finalizer: %w", err)
	}
	return nil
}

// Removes the given finalizer and persists the change.
func RemoveFinalizer(
	ctx context.Context, c client.Client,
	obj client.Object, finalizer string,
) error {
	if !controllerutil.ContainsFinalizer(obj, finalizer) {
		return nil
	}

	controllerutil.RemoveFinalizer(obj, finalizer)

	patch := map[string]interface{}{
		"metadata": map[string]interface{}{
			"resourceVersion": obj.GetResourceVersion(),
			"finalizers":      obj.GetFinalizers(),
		},
	}
	patchJSON, err := json.Marshal(patch)
	if err != nil {
		return fmt.Errorf("marshalling patch to remove finalizer: %w", err)
	}
	if err := c.Patch(ctx, obj, client.RawPatch(types.MergePatchType, patchJSON)); err != nil {
		return fmt.Errorf("removing finalizer: %w", err)
	}
	return nil
}

func EnsureCachedFinalizer(
	ctx context.Context, c client.Client, obj client.Object,
) error {
	return EnsureFinalizer(ctx, c, obj, CachedFinalizer)
}

type cacheFreer interface {
	Free(ctx context.Context, obj client.Object) error
}

// Frees caches and removes the associated finalizer.
func FreeCacheAndRemoveFinalizer(
	ctx context.Context, c client.Client,
	obj client.Object, cache cacheFreer,
) error {
	if err := cache.Free(ctx, obj); err != nil {
		return fmt.Errorf("free cache: %w", err)
	}

	return RemoveFinalizer(ctx, c, obj, CachedFinalizer)
}

type isControllerChecker interface {
	IsController(owner, obj metav1.Object) bool
}

// Returns a list of ControlledObjectReferences controlled by this instance.
func GetControllerOf(
	ctx context.Context, scheme *runtime.Scheme, ownerStrategy isControllerChecker,
	owner client.Object, actualObjects []client.Object,
) ([]corev1alpha1.ControlledObjectReference, error) {
	var controllerOf []corev1alpha1.ControlledObjectReference
	for _, actualObj := range actualObjects {
		if !ownerStrategy.IsController(owner, actualObj) {
			continue
		}

		gvk, err := apiutil.GVKForObject(actualObj, scheme)
		if err != nil {
			return nil, err
		}
		controllerOf = append(controllerOf, corev1alpha1.ControlledObjectReference{
			Kind:      gvk.Kind,
			Group:     gvk.Group,
			Name:      actualObj.GetName(),
			Namespace: actualObj.GetNamespace(),
		})
	}
	return controllerOf, nil
}

// Registers a local field index for the controller referenced in .metadata.ownerReferences.
func RegisterControllerFieldIndex(ctx context.Context, fi client.FieldIndexer, objs ...client.Object) error {
	for _, obj := range objs {
		if err := fi.IndexField(ctx, obj, ControllerIndexFieldKey, indexController); err != nil {
			return err
		}
	}
	return nil
}

func ControllerFieldIndexValue(
	controller client.Object, scheme *runtime.Scheme,
) (string, error) {
	gvk, err := apiutil.GVKForObject(controller, scheme)
	if err != nil {
		return "", err
	}

	key := client.ObjectKeyFromObject(controller)
	return controllerFieldIndexValue(gvk.GroupKind(), key), nil
}

func controllerFieldIndexValue(gk schema.GroupKind, key client.ObjectKey) string {
	return fmt.Sprintf("%s:%s", gk, key)
}

func indexController(o client.Object) []string {
	for _, ref := range o.GetOwnerReferences() {
		if ref.Controller == nil || !*ref.Controller {
			continue
		}

		key := client.ObjectKey{
			Name:      ref.Name,
			Namespace: o.GetNamespace(),
		}
		// ref.Kind
		gv, err := schema.ParseGroupVersion(ref.APIVersion)
		if err != nil {
			panic(err)
		}

		return []string{
			controllerFieldIndexValue(schema.GroupKind{
				Group: gv.Group, Kind: ref.Kind,
			}, key),
		}
	}

	return nil
}
